Some project file reshuffling
