+++ CMakeLists.txt
@@ -147,10 +147,10 @@ ENDIF()
 INCLUDE_DIRECTORIES(${CMAKE_SOURCE_DIR}/${srcdir}/src/ )
 
 # Files
-FILE(GLOB gimagereader_SRCS ${srcdir}/src/*.cc common/*.cc preprocess/*.cpp)
+FILE(GLOB gimagereader_SRCS ${srcdir}/src/*.cc common/*.cc preprocess/*.cc)
 # sort lists to avoid dependency on readdir() order and make build reproducible
 LIST(SORT gimagereader_SRCS)
-FILE(GLOB gimagereader_HDRS ${srcdir}/src/*.hh common/*.hh preprocess/*.hpp)
+FILE(GLOB gimagereader_HDRS ${srcdir}/src/*.hh common/*.hh preprocess/*.hh)
 LIST(SORT gimagereader_HDRS)
 FILE(GLOB gimagereader_FORMS ${srcdir}/data/*.ui)
 LIST(SORT gimagereader_FORMS)
