--- preprocess/Compatibility.hpp
@@ -1,72 +1,0 @@
-/*
-    MIT License
-
-    Copyright (c) 2017 Alexander Zaitsev
-
-    Permission is hereby granted, free of charge, to any person obtaining a copy
-    of this software and associated documentation files (the "Software"), to deal
-    in the Software without restriction, including without limitation the rights
-    to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-    copies of the Software, and to permit persons to whom the Software is
-    furnished to do so, subject to the following conditions:
-
-    The above copyright notice and this permission notice shall be included in all
-    copies or substantial portions of the Software.
-
-    THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-    IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-    FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-    AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-    LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-    OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-    SOFTWARE.
-*/
-
-#ifndef PRLIB_COMPATIBILITY_HPP
-#define PRLIB_COMPATIBILITY_HPP
-
-#include <cctype>
-
-typedef int LONG;
-typedef unsigned char BYTE;
-typedef unsigned int DWORD;
-typedef unsigned short WORD;
-
-struct BITMAPFILEHEADER
-{
-    ushort bfType;
-    uint bfSize;
-    uint bfReserved;
-    uint bfOffBits;
-};
-
-struct BITMAPINFOHEADER
-{
-    uint biSize;
-    int biWidth;
-    int biHeight;
-    short biPlanes;
-    short biBitCount;
-    uint biCompression;
-    uint biSizeImage;
-    int biXPelsPerMeter;
-    int biYPelsPerMeter;
-    uint biClrUsed;
-    uint biClrImportant;
-};
-
-struct RGBQUAD
-{
-    unsigned char rgbBlue;
-    unsigned char rgbGreen;
-    unsigned char rgbRed;
-    unsigned char rgbReserved;
-};
-
-struct BITMAPINFO
-{
-    BITMAPINFOHEADER bmiHeader;
-    RGBQUAD bmiColors[256];
-};
-
-#endif //PRLIB_COMPATIBILITY_HPP
