--- preprocess/ImagePreprocessor.hpp
@@ -1,19 +1,0 @@
-//
-// Created by zamazan4ik on 14.06.17.
-//
-
-#ifndef GIMAGEREADER_IMAGEPREPROCESSOR_HPP
-#define GIMAGEREADER_IMAGEPREPROCESSOR_HPP
-
-#include <QImage>
-
-namespace prl
-{
-class ImagePreprocessor
-{
-public:
-    static QImage preprocess(const QImage& image);
-};
-}
-
-#endif //GIMAGEREADER_IMAGEPREPROCESSOR_HPP
